Extract property and web data stream related error message to separate component.
